requirements.txt
phi==0.6.7
groq==1.3.0
pycoingecko==3.0.0
pandas==1.5.3
python-dotenv==1.0.0
httpx==0.24.0
